requirements.txt
# =============================================================================
# Doctor Appointment Multi-Agent System - Requirements
# =============================================================================

# Core Web Framework
# ------------------
fastapi==0.115.8
uvicorn==0.34.0
streamlit

# AI/ML Framework - LangChain Ecosystem
# -------------------------------------
langchain==0.3.27
langchain-core==0.3.76
langchain-community==0.3.27
langchain-anthropic==0.3.18
langchain-openai==0.3.11
langchain-groq==0.3.8
langgraph==0.2.76
langgraph-checkpoint==2.1.0
langsmith==0.4.13

# Data Processing & Analysis
# --------------------------
pandas==2.2.3
numpy==1.26.4

# Data Validation & Settings
# --------------------------
pydantic==2.10.6
pydantic-settings==2.7.1

# HTTP & Network Communication
# ----------------------------
httpx==0.28.1
requests==2.32.3

# Environment & Configuration
# ---------------------------
python-dotenv==1.0.1

# Logging & Monitoring
# --------------------
loguru==0.7.3 